Created a middleware folder and a Authorize.php containing a class that handles roles coming with requests. also refactored Router to accept midleware

+++ routes.php
@@ -2,20 +2,20 @@
 
 $router->get('/', 'HomeController@index');
 $router->get('/listings', 'ListingsController@index');
-$router->get('/listings/create', 'ListingsController@create');
-$router->get('/listings/edit/{id}', 'ListingsController@edit');
+$router->get('/listings/create', 'ListingsController@create', ['auth']);
+$router->get('/listings/edit/{id}', 'ListingsController@edit', ['auth']);
 $router->get('/listings/{id}', 'ListingsController@show');
 
-$router->put('/listings/{id}', 'ListingsController@update');
-$router->post('/listings/create', 'ListingsController@store');
-$router->delete('listings/{id}', 'ListingsController@destroy');
+$router->put('/listings/{id}', 'ListingsController@update'['auth']);
+$router->post('/listings/create', 'ListingsController@store', ['auth']);
+$router->delete('listings/{id}', 'ListingsController@destroy', ['auth']);
 
-$router->get('/auth/register', 'UserController@create');
-$router->get('/auth/login', 'UserController@login');
+$router->get('/auth/register', 'UserController@create', ['guest']);
+$router->get('/auth/login', 'UserController@login', ['guest']);
 
-$router->post('/auth/register', 'UserController@store');
-$router->post('/auth/logout', 'UserController@logout');
-$router->post('/auth/login', 'UserController@authenticate');
+$router->post('/auth/register', 'UserController@store', ['guest']);
+$router->post('/auth/logout', 'UserController@logout', ['auth']);
+$router->post('/auth/login', 'UserController@authenticate', ['guest']);
 
 /* 
 $router->get('/', 'controllers/home.php');
